Update Weather_Publisher.py
Changed the name of the weather function to main, called main() towards the end of the code to enable running of the function. Additional minor adjustments were made to allow code to be executed, removed unnecessary while True: loop at the end of the code that resulted in constant printing.